% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_utils.R
\name{add_note}
\alias{add_note}
\title{scpo_note}
\usage{
add_note(note, size = 8, width = 132)
}
\arguments{
\item{note}{character string for a note}

\item{size}{font size for the note}

\item{width}{a number of characters to allow before a character return}
}
\value{
a grob formatted for a source in a ggplot
}
\description{
scpo_note
}
