% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_utils.R
\name{add_source}
\alias{add_source}
\title{add_source}
\usage{
add_source(string, size = 8)
}
\arguments{
\item{string}{character string for a source statement}

\item{size}{font size for the source}
}
\value{
a grob formatted for a source in a ggplot
}
\description{
add_source
}
