% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_utils.R
\name{add_subtitle}
\alias{add_subtitle}
\title{add_subtitle}
\usage{
add_subtitle(string, size = 9.5)
}
\arguments{
\item{string}{character string for a subtitle}

\item{size}{font size for the subtitle}
}
\value{
a grob formatted for a subtitle in a ggplot
}
\description{
add_subtitle
}
