% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_utils.R
\name{add_title}
\alias{add_title}
\title{add_title}
\usage{
add_title(string, size = 12, width = 68)
}
\arguments{
\item{string}{character string for a title}

\item{size}{font size for the title}

\item{width}{a number of characters to allow before a character return}
}
\value{
a grob formatted for a source in a ggplot
}
\description{
add_title
}
