% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{bluefocus}
\alias{bluefocus}
\title{Theme labs bluefocus colors palette}
\usage{
bluefocus()
}
\description{
Theme labs bluefocus colors palette
}
\examples{
library(scales)
scales::show_col(bluefocus()(10))
}
