% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_line}
\alias{geom_line}
\title{geom_line in the SciencesPo style}
\usage{
geom_line(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping from ggplot2}

\item{...}{other arguments passed to \code{geom_line()}}

\item{size}{line size}
}
\description{
Submit \code{?ggplot2::geom_line} to see the full documentation for \code{geom_line()}
}
