% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_point}
\alias{geom_point}
\title{geom_point in the SciencesPo style}
\usage{
geom_point(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{mapping from ggplot2}

\item{...}{other arguments passed to \code{geom_point()}}

\item{size}{point size}
}
\description{
Submit \code{?ggplot2::geom_point} to see the full documentation for \code{geom_point()}
}
