% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdecor.R
\docType{package}
\encoding{UTF-8}
\name{ggdecor}
\alias{ggdecor}
\title{\Sexpr[stage=build,results=rd]{tools:::Rd_package_title("ggdecor")}}
\description{
\Sexpr[stage=build,results=rd]{tools:::Rd_package_description("ggdecor")}
}
\references{
Marcelino, Daniel (2016). \emph{ggdecor: A Tool Set for Analyzing Political Behaviour Data}. Available at SSRN: \url{http://dx.doi.org/10.2139/ssrn.2320547}
}
\author{
\Sexpr[stage=build,results=rd]{tools:::Rd_package_author("ggdecor")}
}
