% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{greenfocus}
\alias{greenfocus}
\title{Theme labs greenfocus colors palette}
\usage{
greenfocus()
}
\description{
Theme labs greenfocus colors palette
}
\examples{
library(scales)
scales::show_col(greenfocus()(16))
}
