% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{greenmono}
\alias{greenmono}
\title{Theme labs greenmono 8 colors palette}
\usage{
greenmono()
}
\description{
Theme labs greenmono 8 colors palette
}
\examples{
library(scales)
scales::show_col(greenmono()(8))
}
