% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{font_sans}
\alias{font_sans}
\alias{font_sans_bold}
\alias{font_sans_light}
\alias{font_sans_thin}
\alias{font_sans_humor}
\title{Public Sans regular font name R variable aliases}
\format{
length 1 character vector

length 1 character vector

length 1 character vector

length 1 character vector

length 1 character vector
}
\usage{
font_sans

font_sans_bold

font_sans_light

font_sans_thin

font_sans_humor
}
\description{
\code{font_sans} == "\verb{Public Sans}"

\code{font_sans_bold} == "`Public Sans Bold``"

\code{font_sans_light} == "\verb{Public Sans Light}"

\code{font_sans_thin} == "\verb{Public Sans Thin}"

\code{font_sans_humor} == "\verb{Public Sans Humor}"
}
\note{
\code{font_sans_bold} (a.k.a. "\verb{Public Sans Bold}") is not available on
Windows and will throw a warning if used in plots.
}
\keyword{datasets}
