% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{rainbow10}
\alias{rainbow10}
\title{Theme labs rainbow 10 colors palette}
\usage{
rainbow10()
}
\description{
Theme labs rainbow 10 colors palette
}
\examples{
library(scales)
scales::show_col(rainbow10()(10))
}
