% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{rainbow12}
\alias{rainbow12}
\title{Theme labs rainbow 12 colors palette}
\usage{
rainbow12()
}
\description{
Theme labs rainbow 12 colors palette
}
\examples{
library(scales)
scales::show_col(rainbow12()(12))
}
