% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_utils.R
\name{remove_legend}
\alias{remove_legend}
\title{The SciencesPo [ggplot2] theme}
\usage{
remove_legend(ggplot_object)
}
\arguments{
\item{ggplot_object}{a ggplot}
}
\value{
a ggplot without a legend
}
\description{
Removes legend
}
