% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{font_roboto}
\alias{font_roboto}
\alias{font_roboto_bold}
\alias{font_roboto_light}
\alias{font_roboto_thin}
\title{Roboto regular font name R variable aliases}
\format{
length 1 character vector

length 1 character vector

length 1 character vector

length 1 character vector
}
\usage{
font_roboto

font_roboto_bold

font_roboto_light

font_roboto_thin
}
\description{
\code{font_roboto} == "\verb{Roboto Regular}"

\code{font_roboto_bold} == "\verb{Roboto Bold}"

\code{font_roboto_light} == "\verb{Roboto Light}"

\code{font_roboto_thin} == "\verb{Roboto Thin}"
}
\keyword{datasets}
