% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{scpo_palette}
\alias{scpo_palette}
\title{A muted, qualitative color palette}
\usage{
scpo_palette()
}
\description{
A muted, qualitative color palette
}
\examples{
library(scales)
scales::show_col(scpo_palette()(9))
}
