% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_utils.R
\name{scpo_plot}
\alias{scpo_plot}
\title{scpo_plot}
\usage{
scpo_plot(..., heights = 1)
}
\arguments{
\item{...}{scpo plot objects or grobs}

\item{heights}{relative heights of each object in the final plot}
}
\value{
one plot made from many grobs
}
\description{
scpo_plot
}
