% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_web.R
\name{set_decor_defaults}
\alias{set_decor_defaults}
\title{The SciencesPo \link{ggplot2} theme defaults}
\usage{
set_decor_defaults(style = "web")
}
\arguments{
\item{style}{The default theme style for the R session. "print" or "map".}
}
\description{
\code{set_urbn_defaults} provides a \link{ggplot2} theme formatted according to the
Urban Institute style guide, with sensible defaults.
}
