% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{tableau10_light}
\alias{tableau10_light}
\title{Tableau Classic light 10 colors palette}
\usage{
tableau10_light()
}
\description{
Tableau Classic light 10 colors palette
}
\examples{
library(scales)
scales::show_col(tableau10_light()(10))
}
