% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{tableau10_medium}
\alias{tableau10_medium}
\title{Tableau Classic medium 10 colors palette}
\usage{
tableau10_medium()
}
\description{
Tableau Classic medium 10 colors palette
}
\examples{
library(scales)
scales::show_col(tableau10_medium()(10))
}
