% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{font_xkcd}
\alias{font_xkcd}
\title{xkcd regular font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_xkcd
}
\description{
\code{font_xkcd} == "\code{xkcd}"
}
\keyword{datasets}
